flask-server/requirements.txt
flask
flask_cors
psycopg2-binary
configparser
awsebcli